renaming Logins
