+++ src/components/display/fewlines/AccountOverview/AccountOverview.stories.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { StoriesContainer } from "../StoriesContainer";
 import AccountOverview from "./AccountOverview";
 
-export default { title: "pages/Account", component: AccountOverview };
+export default { title: "pages/AccountOverview", component: AccountOverview };
 
 export const AccountPage = (): JSX.Element => {
   return (
